// packages/flixlix-cards/old-energy-flow-card-plus/src/types.ts
export type ActionConfigExtended = ActionConfig | CallServiceActionConfig | MoreInfoActionConfig;

export interface MoreInfoActionConfig extends BaseActionConfig {
  action: "more-info";
  entity?: string;
  data?: {
    entity_id?: string | [string];
  };
}

export interface CallServiceActionConfig extends BaseActionConfig {
  action: "call-service";
  service: string;
  data?: {
    entity_id?: string | [string];
    [key: string]: unknown;
  };
  service_data?: Record<string, unknown>;
  target?: HassServiceTarget;
  repeat?: number;
  haptic?: HapticType;
}